Implement persistent settings
Settings still doesn't do much, but I've merged HikeManager with it -- it just makes sense for them to be one and the same (considering they both offer up information required throughout the application)

I also made Hike serializable so that we can dump them to disk with the rest of 'em

Final note, MainWindow chucks up an alert box if settings fail to save. The user may as well know.